refactor route handler into separate file, update go.mod for new repo
