+++ main.go
@@ -1,21 +1,10 @@
 package main
 
 import (
-	"github.com/gin-gonic/gin"
+	"github.com/YoshiRussell/bookclubapp/server/routes"
 )
 
-func SetupRouter() *gin.Engine {
-	r := gin.Default()
-	r.GET("/epic", epicHandler)
-	return r
-}
 func main() {
-	router := SetupRouter()
+	router := routes.SetupRouter()
 	router.Run()
-}
-
-func epicHandler(c *gin.Context) {
-	c.JSON(200, gin.H{
-		"message": "fortnite",
-	})
 }
