+++ server/routes/route_handler.go
@@ -1,0 +1,24 @@
+package routes
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+func SetupRouter() *gin.Engine {
+	r := gin.Default()
+	r.GET("/", rootHandler)
+	r.GET("/epic", epicHandler)
+	return r
+}
+
+func rootHandler(c *gin.Context) {
+	c.String(http.StatusOK, "This is a placeholder for the homepage.")
+}
+
+func epicHandler(c *gin.Context) {
+	c.JSON(http.StatusOK, gin.H{
+		"message": "fortnite",
+	})
+}
